refactor(keyword): extract keyword utilities
