+++ src/pages/keyword-generator/KeywordGenerator.vue
@@ -6,6 +6,7 @@ import {
     MaTextarea,
 } from '@mobileaction/action-kit';
 import { computed, ref } from 'vue';
+import { cleanInput, generateUniqueNGrams } from '../../utils/keywordUtils';
 
 const userInput = ref('');
 
@@ -37,15 +38,6 @@ const hasGenerated = computed(() => keywordSections.value
 
 // Require a selection so generation cannot run without a visible result section.
 const hasSelectedGramSizes = computed(() => selectedGramSizes.value.length > 0);
-
-// Normalize the input into lowercase words separated by single spaces.
-const cleanInput = (input) => input
-    .toLowerCase()
-    // \p{L} matches Unicode letters and \p{N} matches Unicode numbers.
-    // This preserves Turkish/non-ASCII text while removing punctuation.
-    .replace(/[^\p{L}\p{N}]+/gu, ' ')
-    .trim()
-    .replace(/\s+/g, ' ');
 
 const cleanedUserInput = computed(() => cleanInput(userInput.value));
 const hasCleanInput = computed(() => Boolean(cleanedUserInput.value));
@@ -88,17 +80,6 @@ const validationMessage = computed(() => {
     // Tell the user the exact threshold needed to produce all selected result groups.
     return `Please provide at least ${largestSelectedGramSize.value} words after unwanted words are removed.`;
 });
-
-// Build unique consecutive word groups while preserving their original order.
-const generateUniqueNGrams = (words, gramSize) => {
-    const keywords = [];
-
-    for (let index = 0; index <= words.length - gramSize; index += 1) {
-        keywords.push(words.slice(index, index + gramSize).join(' '));
-    }
-
-    return [...new Set(keywords)];
-};
 
 // Clean the current input and generate the selected n-gram groups.
 const generateKeywords = () => {
